python-graphene/app/requirements.txt
aniso8601==7.0.0
asgiref==3.2.7
astroid==2.4.1
autopep8==1.5.2
colorama==0.4.3
django-cors-headers==3.2.1
django-graphql-jwt==0.3.1
graphene==2.1.8
graphene-django==2.10.0
graphql-core==2.3.2
graphql-relay==2.0.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
promise==2.3
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.5.2
pytz==2020.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
wrapt==1.12.1
